+++ Editor/EditorConstants.cs
@@ -1,0 +1,11 @@
+
+
+namespace SimpleMan.Utilities.Editor
+{
+    internal static class EditorConstants
+    {
+        //------FIELDS
+        public static string MAIN_PACKAGE_PATH = "Assets/Plugins/SimpleMan/Utilities/Editor/Installer/SimpleMan_Utilities_v1.1.5.unitypackage";
+        public static string ODIN_INSPECTOR_PATH = "Assets/Plugins/Sirenix/Odin Inspector";
+    }
+}